# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Olyxee Enterprise Logistics Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /internal/voice/order-lookup:
    post:
      operationId: internalVoiceOrderLookup
      tags: [internal-voice]
      summary: Look up an order for the Retell voice agent
      description: |
        Server-to-server endpoint called by Retell during a live call.
        Authenticated by RETELL_INTERNAL_TOKEN bearer header.
      security:
        - internalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderLookupInput"
      responses:
        "200":
          description: Order lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderLookupResult"
        "401":
          description: Invalid internal token
        "404":
          description: Business not found for this number
  /internal/voice/request-human:
    post:
      operationId: internalVoiceRequestHuman
      tags: [internal-voice]
      summary: Record a request-human escalation from the Retell voice agent
      description: |
        Server-to-server endpoint called by Retell when the agent triggers
        a request_human tool call.
      security:
        - internalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestHumanInput"
      responses:
        "200":
          description: Escalation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestHumanResult"
        "401":
          description: Invalid internal token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    RequestHumanResult:
      type: object
      properties:
        received:
          type: boolean
...
    OrderLookupInput:
      type: object
      properties:
        call_id:
          type: string
        from_number:
          type: string
        to_number:
          type: string
        tracking_id:
          type: string
        attempt:
          type: number
    OrderLookupResult:
      type: object
      properties:
        found:
          type: boolean
        call_id:
          type: string
        attempt:
          type: number
        order:
          $ref: "#/components/schemas/OrderLookupOrder"
    OrderLookupOrder:
      type: object
      properties:
        tracking_id:
          type: string
        reference:
          type: string
          nullable: true
        order_reference:
          type: string
          nullable: true
        current_status:
          type: string
        status:
          type: string
        status_label:
          type: string
        statusLabel:
          type: string
        estimated_delivery_date:
          type: string
          nullable: true
        last_updated:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/OrderLookupEvent"
    OrderLookupEvent:
      type: object
      properties:
        at:
          type: string
        timestamp:
          type: string
        status:
          type: string
        label:
          type: string
        status_label:
          type: string
        statusLabel:
          type: string
        message:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
    RequestHumanInput:
      type: object
      properties:
        call_id:
          type: string
        reason:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    internalToken:
      type: http
      scheme: bearer